ca_desktop/backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
python-jose[cryptography]>=3.3.0
pyjwt>=2.8.0
passlib>=1.7.4
bcrypt>=4.0.0
python-multipart>=0.0.6
pydantic[email]>=2.10.0
pydantic-settings>=2.1.0
aiofiles>=23.2.1

# Email service
resend>=0.8.0

# QR code generation
qrcode[pil]>=7.4.2
pillow>=10.0.0

# Slug generation
python-slugify>=8.0.1

# Excel file parsing
openpyxl==3.1.2
pandas>=2.2.0

# Database migrations
alembic>=1.13.0

# Testing
pytest>=7.4.3
httpx>=0.25.1
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
